crewai_research/requirements.txt
crewai==0.41.1
crewai-tools==0.8.3
python-dotenv==1.0.0
openai==1.35.3
langchain==0.2.5
langchain-openai==0.1.8